+++ scripts/handle_plugin_preferences.js
@@ -3,10 +3,7 @@
 var helpers = require('./helpers');
 
 module.exports = function (context) {
-    var path = context.requireCordovaModule('path');
     var fs = context.requireCordovaModule('fs');
-    var cordovaUtil = context.requireCordovaModule('cordova-lib/src/cordova/util');
-    var projectRoot = cordovaUtil.isCordova();
     var callbackPreference = getPreference('ENABLE_CORDOVA_CALLBACK');
     var debugPreference = getPreference('ENABLE_DEBUG_DETECTION');
 
@@ -22,10 +19,8 @@ module.exports = function (context) {
     }
 
     helpers.getPlatformsList(context).forEach(function (platform) {
-        var platformPath = path.join(projectRoot, 'platforms', platform);
-        var pluginDir;
-        var sourceFile;
-        var content;
+        var source = helpers.getPluginSource(context, platform);
+        var content = source.content;
 
         if (callbackPreference) {
             process.stdout.write('[ANTI-TAMPERING] Enabling cordova callbacks for ' + platform + '\n');
@@ -35,14 +30,6 @@ module.exports = function (context) {
         }
 
         if (platform === 'android') {
-            pluginDir = path.join(platformPath, 'src');
-            sourceFile = path.join(pluginDir, 'com/duddu/antitampering/AntiTamperingPlugin.java');
-            try {
-                content = fs.readFileSync(sourceFile, 'utf-8');
-            } catch (e) {
-                exit('Unable to read java class source at path ' + sourceFile, e);
-            }
-
             if (callbackPreference) {
                 content = content.replace(/.*checkAndStopExecution\(\);[\r\n]*/, '');
             }
@@ -51,23 +38,13 @@ module.exports = function (context) {
             }
 
             try {
-                fs.writeFileSync(sourceFile, content, 'utf-8');
+                fs.writeFileSync(source.path, content, 'utf-8');
             } catch (e) {
-                exit('Unable to write java class source at path ' + sourceFile, e);
+                helpers.exit('Unable to write java class source at path ' + source.path, e);
             }
         }
 
         if (platform === 'ios') {
-            var IosParser = context.requireCordovaModule('cordova-lib/src/cordova/metadata/ios_parser');
-            var iosParser = new IosParser(platformPath);
-            pluginDir = path.join(iosParser.cordovaproj, 'Plugins', context.opts.plugin.id);
-            sourceFile = path.join(pluginDir, 'AntiTamperingPlugin.m');
-            try {
-                content = fs.readFileSync(sourceFile, 'utf-8');
-            } catch (e) {
-                exit('Unable to read obj-c source at path ' + sourceFile, e);
-            }
-
             if (callbackPreference) {
                 content = content.replace(/.*\[self checkAndStopExecution\];[\r\n]*/, '');
             }
@@ -76,15 +53,10 @@ module.exports = function (context) {
             }
 
             try {
-                fs.writeFileSync(sourceFile, content, 'utf-8');
+                fs.writeFileSync(source.path, content, 'utf-8');
             } catch (e) {
-                exit('Unable to write obj c source at path ' + sourceFile, e);
+                helpers.exit('Unable to write obj-c source at path ' + source.path, e);
             }
         }
     });
-
-    function exit (msg, exception) {
-        process.stdout.write('\n[ANTI-TAMPERING] ERROR! ' + msg + '\n');
-        throw new Error(exception);
-    }
 };