+++ scripts/helpers/error_exit.js
@@ -1,0 +1,4 @@
+module.exports = function (msg, exception) {
+    process.stdout.write('\n[ANTI-TAMPERING] ERROR! ' + msg + '\n');
+    throw new Error(exception);
+};